+++ MakgeolliGuru/app/build.gradle
@@ -6,14 +6,15 @@ plugins {
 }
 
 android {
-    namespace 'com.example.makgeolliguru'
+    namespace 'com.myapp.makgeolliguru'
     compileSdk 35
 
     defaultConfig {
-        applicationId "com.example.makgeolliguru"
-        minSdk 25
-        versionCode 1
-        versionName "1.0"
+        applicationId "com.myapp.makgeolliguru"
+        targetSdkVersion 34
+        minSdk 28
+        versionCode 4
+        versionName "1.0.3"
 
         testInstrumentationRunner "androidx.test.runner.AndroidJUnitRunner"
     }
@@ -44,11 +45,9 @@ dependencies {
     implementation 'androidx.compose.material3:material3-android:1.3.1'
     implementation 'androidx.work:work-runtime-ktx:2.10.0'
     testImplementation 'junit:junit:4.13.2'
-    implementation platform('com.google.firebase:firebase-bom:33.9.0')
-    implementation 'com.google.firebase:firebase-analytics'
 
-    implementation "com.google.android.gms:play-services-measurement-api:22.2.0"
-    implementation "com.google.android.gms:play-services-measurement-impl:22.2.0"
+
+
 
     implementation 'com.squareup.retrofit2:retrofit:2.9.0'
     implementation 'com.squareup.retrofit2:converter-gson:2.9.0'
