+++ MakgeolliGuru/app/google-services.json
@@ -9,7 +9,7 @@
       "client_info": {
         "mobilesdk_app_id": "1:658914760316:android:3805a47a24cd9b86e25bf9",
         "android_client_info": {
-          "package_name": "com.example.makgeolliguru"
+          "package_name": "com.myapp.makgeolliguru"
         }
       },
       "oauth_client": [],
