+++ MakgeolliGuru/app/src/main/AndroidManifest.xml
@@ -34,5 +34,9 @@
     <uses-permission android:name="android.permission.WRITE_EXTERNAL_STORAGE" />
     <uses-permission android:name="android.permission.READ_EXTERNAL_STORAGE" />
 
+    <meta-data
+        android:name="google_analytics_adid_collection_enabled"
+        android:value="false" />
+
 
 </manifest>
