--- MakgeolliGuru/app/src/main/java/com/example/makgeolliguru/articles/Article.kt
@@ -1,6 +1,0 @@
-package com.example.makgeolliguru.articles
-
-data class Article (
-    val title: String,
-    val imageUrl: String
-)
