--- MakgeolliGuru/app/src/main/java/com/example/makgeolliguru/profile/Favorite.kt
@@ -1,8 +1,0 @@
-package com.example.makgeolliguru.profile
-
-data class Favorite(
-    val name: String,
-    val imageUrl: String,
-    val localisation: String,
-    val data: Array<String>?
-)
