+++ MakgeolliGuru/app/src/main/java/com/myapp/makgeolliguru/articles/Article.kt
@@ -1,0 +1,6 @@
+package com.myapp.makgeolliguru.articles
+
+data class Article (
+    val title: String,
+    val imageUrl: String
+)
