+++ MakgeolliGuru/app/src/main/java/com/myapp/makgeolliguru/profile/Favorite.kt
@@ -1,0 +1,8 @@
+package com.myapp.makgeolliguru.profile
+
+data class Favorite(
+    val name: String,
+    val imageUrl: String,
+    val localisation: String,
+    val data: Array<String>?
+)
